api-python/requirements.txt
azure-functions
azure-data-tables
azure-storage-blob
PyJWT
bcrypt
openai
python-dotenv
pymssql
openpyxl
python-multipart
numpy
pandas
scipy
